chore: add playground (see commit message)
This contains a very rudementary playground for testing components. It
is not complete, nor has it been finalised. It still needs work to
become a solid example for pure CSS implementation. Until then, you can
only use it as a roughe example of implementation. Please refer to the
proper documentation for Cupid to implement this properly.